fix: allow password change when allow_password_change is true
When a user initiates the password reset process, allow_password_change
is set to true.
At this time, if current_password is submitted, update is executed,
resulting in an UnknownAttributeError.
To avoid this error, I modified the code to execute update_with_password
even when allow_password_change is true if params include
current_password.

Refs: #185