feat: include the current user's contact info in the user's details
Edit the `show` action in `UsersController` to set the user instance variable and include the current user's contact information in the response. Introduce a private method to retrieve the current user's contact for the specified user. Update the `UserResource` to conditionally include the current user contact in the serialized response.